backend/requirements.txt
python-dotenv==1.2.2
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.5 
mysqlclient==2.2.8
pydantic==2.5.2
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.23.2